requirements.txt
streamlit>=1.28.0
openai>=1.3.0
PyPDF2>=3.0.0
python-docx>=1.0.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tiktoken>=0.5.1
numpy>=1.24.0
pandas>=2.0.0
